Fixed bug in EventIndex & in all Index views
Data in EventIndex wasn't being reloaded when the component was called from a different route.
Fixing the mentioned bug led to another bug being discovered in all of the Index views. Information from the event emitted by Pagination was not being propperly handled in the parent component.
The latter has been fixed.